fix: move test controller to items
